lstat hook
