+++ sysdeps/nacl/lxstat.c
@@ -3,8 +3,23 @@
 
 int __lxstat (int vers, const char *name, struct stat *buf)
 {
-  errno = ENOSYS;
-  return -1;
+  if (buf == NULL || path == NULL)
+    {
+      errno = EFAULT;
+      return -1;
+    }
+  struct nacl_abi_stat st;
+  int result = __nacl_irt_lstat (path, &st);
+  if (result != 0)
+    {
+      errno = result;
+      return -1;
+    }
+  else
+    {
+      __nacl_abi_stat_to_stat (&st, buf);
+      return 0;
+    }
 }
 hidden_def (__lxstat)
 weak_alias (__lxstat, _lxstat);
